Refactored the whole code with new code architecture
